Advanced Projects/self-learning_chatbot/self_learning_support_chatbot/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
langchain
pydantic
langchain-openai
langchain-community


